+++ base/auth.py
@@ -6,7 +6,8 @@ from flask import (redirect,
                    url_for,
                    session,
                    request,
-                   flash)
+                   flash,
+                   Markup)
 from base.application import app
 from functools import wraps
 from base.models2 import user_m
@@ -17,15 +18,21 @@ from logzero import logger
 from flask import Flask, redirect, url_for
 from flask_dance.contrib.google import make_google_blueprint, google
 from flask_dance.contrib.github import make_github_blueprint, github
+#from flask_dance.contrib.dropbox import make_dropbox_blueprint, dropbox
 from flask_dance.consumer import oauth_authorized
 
-google_bp = make_google_blueprint(scope=["https://www.googleapis.com/auth/userinfo.profile", "https://www.googleapis.com/auth/userinfo.email"])
-github_bp = make_github_blueprint()
+
+google_bp = make_google_blueprint(scope=["https://www.googleapis.com/auth/userinfo.profile", "https://www.googleapis.com/auth/userinfo.email"],
+                                  offline=True)
+github_bp = make_github_blueprint(scope="user:email")
+# dropbox_bp = make_dropbox_blueprint()
 
 
 @app.route("/login/select", methods=['GET'])
 def choose_login(error=None):
     # Relax scope for Google
+    if not session.get("login_referrer", "").endswith("/login/select"):
+        session["login_referrer"] = request.referrer
     os.environ['OAUTHLIB_RELAX_TOKEN_SCOPE'] = "true"
     VARS = {'page_title': 'Choose Login'}
     if error:
@@ -37,23 +44,30 @@ def choose_login(error=None):
 def authorized(blueprint, token):
     if google.authorized:
         user_info = google.get("/oauth2/v2/userinfo")
+        logger.debug(user_info)
+        assert user_info.ok
         user_info = {'google': user_info.json()}
         user_email = user_info['google']['email'].lower()
     elif github.authorized:
+        user_emails = github.get("/user/emails")
+        user_email = [x for x in user_emails.json() if x['primary']][0]["email"].lower()
         user_info = {'github': github.get('/user').json()}
-        logger.info(user_info['github'])
-        if not user_info['github'].get('email'):
-            flash('Please set a public email address in github', 'error')
-            return redirect(url_for('choose_login'))
-        user_email = user_info['github']['email'].lower()
+        user_info['github']['email'] = user_email
+    # elif dropbox.authorized:
+    #     logger.debug(dropbox.authorized)
+    #     dbx_info = dropbox.get("/get_account").json()
+    #     logger.debug(dbx_info)
+    #     user_email = dbx_info["email"]
+    #     user_info = {'dropbox': dbx_info}
     else:
-
         flash("Error logging in!")
         return redirect(url_for("choose_login"))
 
-    #Create or get existing user.
+    # Create or get existing user.
     logger.info(user_email)
     user = user_m(user_email)
+    logger.debug(user)
+    logger.debug(user._exists)
     if not user._exists:
         user.user_email = user_email
         user.user_info = user_info
@@ -62,13 +76,14 @@ def authorized(blueprint, token):
         user.username = slugify("{}_{}".format(user_email.split("@")[0], unique_id()[0:4]))
 
     user.last_login = arrow.utcnow().datetime
+    logger.debug(user)
     user.save()
 
     session['user'] = user
-    logger.info(session)
+    logger.debug(session)
 
     flash("Successfully logged in!", 'success')
-    return redirect(url_for('primary.primary'))
+    return redirect(session.get("login_referrer", None) or url_for('primary.primary'))
 
 def login_required(f):
     @wraps(f)
@@ -86,6 +101,5 @@ def logout():
         Logs the user out.
     """
     session.clear()
-    print(session)
     flash("Successfully logged out", "success")
     return redirect(request.referrer)