+++ base/views/primary.py
@@ -27,7 +27,7 @@ def primary():
     files = sorted_files("base/static/content/news/")
     VARS = {'page_title': page_title,
             'files': files,
-            'latest_mappings': get_latest_public_mappings()}
+            'latest_mappings': []}
     return render_template('primary/home.html', **VARS)